initial draft of daq class. moved linpots to daq class
