+++ DAQ_System/Source/main.cpp
@@ -13,8 +13,7 @@
 #include "mbed.h"
 
 // DFR Custom Dependancies
-#include "Adapter/Interfaces/ilinear_potentiometer.hpp"
-#include "platform/component_interface_bridge.hpp"
+#include "Platform/daq.hpp"
 
 using AutoReloadTimer = LowPowerTicker;
 
@@ -23,22 +22,12 @@ static void start_logging() {
     logging = true;
 }
 
-struct SuspensionPotentiometers {
-    std::unique_ptr<adapter::ILinear_Potentiometer> front_left;
-    std::unique_ptr<adapter::ILinear_Potentiometer> front_right;
-    std::unique_ptr<adapter::ILinear_Potentiometer> rear_left;
-    std::unique_ptr<adapter::ILinear_Potentiometer> rear_right;
-};
+
 
 int main() {
     // Init components
-    platform::ComponentInterfaceBridge bridge;
-
-    SuspensionPotentiometers suspension_pots;
-    suspension_pots.front_left  = bridge.GetLinearPotentiometer(platform::front_left);
-    suspension_pots.front_right = bridge.GetLinearPotentiometer(platform::front_right);
-    suspension_pots.rear_left   = bridge.GetLinearPotentiometer(platform::rear_left);
-    suspension_pots.rear_right  = bridge.GetLinearPotentiometer(platform::rear_right);
+    platform::DAQ daq;
+    daq.Init();
     
     // Start timer
     constexpr uint8_t kLoggingRate = 3; // seconds
@@ -49,13 +38,13 @@ int main() {
 
     // Operate
     while (true) {
-        suspension_pots.front_left->ComputeDisplacementPercentage();
+        daq.suspension_pots.front_left->ComputeDisplacementPercentage();
 
         if (logging) {
             // Operate: Writing
             timestamp += kLoggingRate;
-            std::cout << suspension_pots.front_left->GetDisplacementInches() << " in\t"
-                      << suspension_pots.front_left->GetDisplacementMillimeters() << " mm" << std::endl;
+            std::cout << daq.suspension_pots.front_left->GetDisplacementInches() << " in\t"
+                      << daq.suspension_pots.front_left->GetDisplacementMillimeters() << " mm" << std::endl;
             
             logging = false;
         }
